+++ conreality_fonts/lib/conreality_fonts.dart
@@ -1,5 +1,6 @@
 /* This is free and unencumbered software released into the public domain. */
 
+/// Font assets (B612 and B612 Mono) for Conreality games.
 library conreality_fonts;
 
 export 'b612.dart' show B612;