+++ conreality_fonts/CHANGELOG.md
@@ -6,5 +6,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.0.1 - 2019-07-02
+## 1.0.0 - 2019-07-03
 ### Added
+- `B612.version` getter
+- `B612.fontFamily` getter
+- `B612.normal` getter
+- `B612.italic` getter
+- `B612.bold` getter
+- `B612.boldItalic` getter
+- `B612Mono.version` getter
+- `B612Mono.fontFamily` getter
+- `B612Mono.normal` getter
+- `B612Mono.italic` getter
+- `B612Mono.bold` getter
+- `B612Mono.boldItalic` getter